bus: introduce ReceiveBuffer
This is mostly behavior preserving refactoring which extracts "parsing"
related  parts of MessageBus into a sans-io ReceiveBuffer. As a result,
now the replica the drives parsing loop, allowing for batch processing
(prioritization & selective back-pressure), although the actual changes
to processing are left for the follow-up.

This seems like a good refactor anyway, as it allows us to fuzz this
component easily.