ci: speedup lsm_scan_fuzz
It runs out of 30min budget on CFO.

The fuzzer is slow because it is quadratic worst-case: every commit adds
a number of items to the database, and each query is linear in the
database.

I don't want to reduce the number of commits, as we want to fill LSM
levels. So instead I just reduce the frequency of queries: we'll still
be probabilistically calling every query at every point!

The actual formula to compute query probability is hairy, and I stopped
understanding it the moment I typed the last closing `)`. I do trust the
comptime tests though that it is not always returning zero!

Seed: ./zig/zig build -Drelease fuzz -- lsm_scan 1112800611640318623