bus: selective back pressure
Before this commit, replica had to process every message that the bus
throws at replica. Now, replica gains the ability to "suspend" certain
messages in bus' receive buffer, to return to them later.

This is used to:

* postpone prepare messages until there's a journal write available
* postpone block messages until there's a grid repair write available

If a message is suspended, the bus won't be recving into the socket,
propagating backpressure all the way up to the sending replica.

To ameliorate head-of-line blocking issues, backpressure applies to
specific commands. If there are _many_ messages in a particular buffer,
they all will be considered.

If a replica suspends a message, it must call bus.ready_to_receive at
some later point to guarantee that messages are delivered.