vopr: assert that replicas agree on commits
StateChecker tracks the sequence of commits made logically by the
cluster, by observing individual's replicas commits and advancing the
watermark.

Unless I am mistaken, if the water mark is not advanced (that is, a
different replcia committed first), we _don't_ check that we commit the
same prepare. This feels like an oversight.