stdx: refine BoundedArrayType API
Rename append_assume_capacity to push, to be shorter and consistent with
Queue&Stack.

Also make it return `?void`, so that the caller has to `.?` explicitly,
signaling a pre-condition here.

I am a bit torn on void+callee assert over void?+caller .?, but I prefer
either to `assume_capacity`, and doing from caller to callee is the
easier future move.