vsr: cleanup, handle message padding uniformly
Message that is transferred as n bytes on the wire might take up to
n + (sector_padding - 1) when written to disk, because disks are block
devices operating on sector granularity.

For most messages we used to fix this in the message bus, but for
replies the fix was applied right before writing it to disk.

Now, all the message types are handled uniformly by the Replica's
`on_messages`, where we can also guarantee that message ref-count is
zero, and so it is valid to mutate the underlying buffer.