vopr: specify io-delay in nanoseconds
This is the third installment in the tickless VOPR series, whose overall
goal is to make sure that the VOPR can:

* simulate IO that completes faster than in a tick
* arbitrary permute io requests around each other

Historically, VOPR specified IO delay in terms of _ticks_, so the
"tickless" logic required writing a fancy priority queue which could
pick an arbitrary event, if many events are ready at the same tick.

But that's a very round-about way to solve the problem! A more direct
solution is to model latencies in nanoseconds! There are _many_
nanoseconds, so a _random_ timestamp now properly shuffles events!

A side benefit here is that, for things like performance VOPR, its now
much easier to map simulated latencies to real-world equivalents.