fuzz: fix stack probing to not actually _use_ all the stack
https://github.com/tigerbeetle/tigerbeetle/pull/2999 added eager stack
probing to our fuzz runs. The probing worked correctly in the sense that
it did allocate 3 MiBs of stack and touched it.

What didn't work though, was that the probe itself was inlined, and so
the stack  space was release only after return from main. That is, we
didn't just _probe_ the stack, but actually used all of it!

This of course is visible in the original godbolt: https://godbolt.org/z/Ez9r69rW6

Marking the function as noinline achieves the desired result where we
just touch the stack, and then give it back to the actual fuzzing code.