replica: add primary and status information to all log lines
When glancing at server logs, it's really helpful to know which replica is
the primary. This can already be figured out by knowing the topology and
the view number, but this makes it much quicker.

Additionally, include the status if it's not `.normal`.

(I briefly considered making replica have a `self.log.info` & co, but figured
this was more in keeping with the existing spirit...)