benchmark: print only key latencies
Latency histogram is the key output of the benchmark, and is something
directly processed by humans. It's worth investing into increasing
signal to noise ratio here, and to print _less_ data, such that the data
that are printed become more valuable. p1, p50, p99 and p100 feel like
the key numbers for us.