vsr: fix a crash
realtime clock is not guaranteed to be  monotonic and can regress
in-process. Furthermore, prepare timestamp might be backstopped from
bellow by the clock of the _previous_ primary, so this assert can fire
even when time _is_ monotonic in process.

In general, I am not happy how we are mixing up logical&physical time
here, but I want to look into that separately.

Seed: ./zig/zig build vopr -- --lite 12439570601600962653