cfo: fix weight
We have a very sophisticated algorithm for relatively weighing different
fuzzers, whose goal is to allocate CPU time (rather than just number of
runs) proportional to fuzzer's importance.

But, due to a bug, the scheduler ended up being completely fair, just
picking a fuzzer uniformly.

The bug is that our weights are largish (million / #fuzzers), and we
divide 1000 by weight to get virtual runtime. But with large weights
that's just always zero.

Fix by bumping the constant to 10**9. We use u64, so this shouldn't
overflow.