VSR: Add metric measuring client request round trip latency
The client tracks both the instant and the wall clock time -- the client's instant is used for the duration, but is not meaningful at the replicas, so we use wall clock for the timestamp.

This adds way more metrics "worst-case" packets than it reasonably should, but only because the worst-case calculation assumes that there are >140 different operations.

Currently the previous request timestamp is not used by the replica.