VSR/Client: Measure latency from earlier
The latency begins from on the first event is added to the multi batch, even before the request is ever sent. That is, it includes the time that the request spends queued up on the client. The goal is to get an accurate representation of how much latency the application actually observes.