vsr: remove dead code
Due to

    prepare.ok_from_all_replicas.count() >= self.quorum_replication)

and

    if (prepare_pipelined.ok_from_all_replicas.count() >= self.quorum_replication) {
        prepare_pipelined.ok_quorum_received = true;
    }

checks that dominate the code in question, prepare can't be pending
anymore, so `==` is tautologically false.

I changed the code to what it does (reset), rather that to what it
intended to do (backoff), because reset seems more appropriate here, as
we did in fact make progress.

See also #2813