VSR: Apply backpressure when replicas are lagging
Apply backpressure back to the client (via slowing down the primary's commit pipeline) when one or more replicas are lagging behind.

This is currently a bit of a kludge (several magic numbers). Experimentally it reduces the chance that replicas in normal operation lag into state sync. When they do lag into state sync, they are able to complete sync + repair.