VSR: Apply backpressure when replicas are lagging
Apply backpressure back to the client (via slowing down the primary's commit pipeline) when one or more replicas are lagging behind.

This is currently a bit of a kludge (several magic numbers). Experimentally it reduces the chance that replicas in normal operation lag into state sync. When they do lag into state sync, they are able to complete sync + repair.

The stall is positioned in the commit pipeline since:
- we need the pipeline `Prepare` to get the acks, and
- `reply_setup` must synchronously proceed the execution stage.

The increment to `packet_count_max` is due to the new commit stage, which (slightly) increases the number of metrics.