compaction: always write full value/index blocks at the end of the beat
There is a non-determinism bug that we could hit if we don't write full
value blocks at the end of the beat. Specifically, if based on the spare
capacity of output_blocks, different replicas may choose to either write
or not write a value block at the end of the beat. If one replica writes
it at the end of the beat, but another doesn't, the other replica could
assign a different address to the same value block (since ManifestLog
grid reservation for the next beat could acquire that block).