tests: generate unit_tests.zig
Inspired by https://github.com/tigerbeetle/tigerbeetle/pull/3124#discussion_r2246103103
this generates the `unit_tests.zig` file in `build.zig`.

There's a bit of logic as to what files should be included, so I thought it's
more evident to go with a generated in-repo file rather than removing
`unit_tests.zig` entirely.

Another option could be to keep `unit_tests.zig` as manually generated but lint
it to ensure everything with a `test` block is included.