VSR: `send_request_blocks()` via `repair()`
This is primarily important during the initial phase of state sync, when we sync the client sessions and manifest trailers.

Requesting grid blocks is triggered by receiving grid blocks -- when we receive a block that we need, we restore budget and if we have enough budget we make a new request.

This works great during table sync, but during trailer sync it doesn't work well. The trailers are repaired sequentially (free set → client sessions → manifest) but between each of these we need to do a read (to check for the grid block on our disk). That means that e.g. at the moment we repair the last free set block, we restore budget, then check whether we have more blocks to repair, but we don't yet. In a moment the client sessions read finishes (fails) and we queue a block for repair, but it is too late. So we wait for `grid_repair_message_timeout` (~500ms) before triggering the next request. Then this happens again during the transition from client sessions to manifest. This adds about a second of latency to state sync!

To fix, we now send the block request from `repair()`. `repair()` is on a more frequent timer, but also it is called proactively often.
The `grid_repair_message_timeout` is still used to restore budget.

Some other possible fixes:
1. Concurrently repair all the trailers.
2. Send a summary of trailer blocks in the SV.
3. Add a new "survey" block which includes the address/checksums of all trailer blocks.
4. Wire up an event from grid → replica when a block is queued for repair.
5. Reduce `grid_repair_message_timeout`.

Probably we should eventually implement 1 and 2-or-3, but I think this change is useful even then, so that we are a little more proactive about repair.

I wanted to avoid 4, to avoid coupling replica + grid together in that direction.
5 only reduces the scale of the problem, but doesn't actually solve it.