cli: don't remove datafile if formatting fail
Formatting is atomic, by virtue of superblock writes being atomic. If
something fails, its better to let the user do the cleanup, cleaning up
automatically might mess things up, and is pretty hard to test.

In particular, we are actually using unlinkat wrong here, as we pass
parent directory as a fd, but we use path relative to cwd!