cli: don't unlink data file on formatting failure
Formatting is atomic, as writing the superblock is atomic, so there's no
_need_ to delete. And not deleting helps to avoid hard-to-test-for bugs,
like passing a wrong base directory to unlinkat :o)

There's an argument to be made that unlinking makes sense for reformat
flow, where we do

    open_data_file()
    view = use_client_to_get_safe_view();
    format(view)

and there's a rather sizable window where the data file exists, but is
not formatted. One would think that this could be fixed with this
rearrangement,

    view = use_client_to_get_safe_view();
    open_data_file()
    format(view)

but that has another drawback --- we might fail to create a file for a
stupid reason (mistyped directory name), and its better to learn about
that before starting the whole reformat protocol dance.

But, on balance, just leaving the non-formatted file in-place seems like
the best solution: while moderately less convenient in an edge case, it
removes a hard-to-test and dangerous operation!