VSR: Lift `request_blocks` construction into replica
This is setup for the following commit, which will track in-flight block requests as part of the `RepairBudgetGrid`.

But also, this code was previously part of `Grid`/`GridBlocksMissing` because it manipulated their state internally as a workaround for the lack of proper in-flight request checking.

`GridBlocksMissing` still has `faulty_blocks_repair_index` but this is just an optimization. In some cases we may still need to iterate the whole `faulty_blocks` to find a block to request, since removals are via `swapRemoveAt`.