multiversion: avoid untrodden paths during upgrade
During upgrades, tigerbeetle dynamically re-executed itself at a different version.
There can be only one tigerbeetle instance running at a time, due to any of:
- data file lock,
- incoming port,
- available RAM.
On POSIX, execve semantics of replacing the parent process gives us this for free.
On Windows, parent and child cooperate, with the child blocking until the parent exit.
Blocking is achieved by passing down an anonymous pipe, a-la
<https://matklad.github.io/2023/10/11/unix-structured-concurrency.html>

The previous implementation attempted to achieve a similar effect by
de-initing the replica in the data file but that's actually panic prone
(storage asserts, on deinit, that there are no next ticks scheduled),
and, as a general rule, shutdown code that runs only on one operating
system only during upgrade is bound to become buggy sooner or later.