VSR: State sync ratchet
## Context

State sync overview:

1. Replica discovers that it is lagging behind the cluster.
2. Lagging replica replaces its superblock (in particular the `CheckpointState`) with a newer one from the cluster.
3. Lagging replica syncs the trailer blocks referenced by the new superblock.
4. Lagging replica syncs the table index/value blocks referenced by the new manifest.

In a large database, step 4 might involve syncing a lot of data, and so take quite a long time.

And during this whole time, the remainder of the cluster is likely still processing requests and periodically checkpointing.

Previously, each time we change the sync target checkpoint, we start sync over from the beginning. When we get to table sync, we potentially need to re-sync tables that were already synced.

In this PR, we ratchet progress of table sync, to retain progress across checkpoint boundaries.

## Fix

In `Replica`:
- Add `sync_tables_op_range`.
- If we load a new sync target, we continue syncing the old range already in `sync_tables_op_range` until it is complete, at which point we change the range to cover all unsynced ops.

In `GridBlocksMissing`:
- When we begin the replace the superblock for state sync, we pause table sync. (See the `faulty_blocks`/`syncing_faulty_blocks` swap.)
- When we open that's grid/forest, we resume table sync. If any of the tables we were syncing didn't survive into the new checkpoint, they are canceled.

Note that if a replica restarts before persisting the fact that it completed state sync (via a checkpoint) then it will still need to start sync over.