References the `native` subfolder so external files required by CGO aren't pruned.
We could remove the `native` folder and move all files to the root.
See ["#including source files in subdirectories is a bad idea"](https://go-review.googlesource.com/c/go/+/125297/5/src/cmd/cgo/doc.go).

But we prefer to keep the root clean and minimal, so we embrace the `native` subfolder and declare a dummy package there to force references.

Additionally, the compiled library carries the platform as part of the filename instead of using subfolders, avoiding the need to declare separate packages for each platform.