CFO: Boost release branch weight during ~weekend
Boost fuzzing of release branch for 2 days after the release branch is committed to. In the usual case, we will push to release branch on Friday, so it will be allocated bonus fuzzing time on Saturday/Sunday.

The 3x weight multiplier was chosen arbitrarily.

This was inspired by matklad's original suggestion, to boost weight just during the weekend. But I think this approach is a bit more flexible -- we probably would also want to boost if e.g. we need to release an off-schedule hotfix.

I tested this commit on a CFO server to make sure that it didn't break anything. But we'll need to wait till Friday to see it in action.