logs: add --log-trace and gate per event + tracing logs
These two logs, per event in the state machine and tracing output account for
the vast majority of logs in debug mode.

By gating them behind a new flag, `--log-trace`, running with `--log-debug`
is now fast enough to be enabled in production!

Done as an explicit parameter and check, rather than overriding the entire
logging library, or adding a new function and a global: I'm not sure we
want to make adding new cases like this too easy!