multiversion: avoid untrodden paths during upgrade
During upgrades, tigerbeetle dynamically re-executes itself at a different version.
There can be only one tigerbeetle instance running at a time, due to any of:
- data file lock,
- incoming port,
- available RAM.
On POSIX, execve semantics of replacing the parent process gives us this for free.
On Windows, parent and child cooperate, with the child blocking until the parent exit.
Specifically:
1. Parent creates a named pipe with a random name.
2. Parent passes the name of this pipe to the child via inherited env variable.
3. Parent duplicates its own handle into the child process.
4. Parent passes the value of this handle to the child via anonymous pipe
   (We don't want to pass parent handle _directly_ to avoid handle inheritance)
5. Child gets the name of the pipe trough env.
6. Child receives the parent's handle through the pipe.
7. Child waits for parent to exit.

The previous implementation attempted to achieve a similar effect by
de-initing the replica in the data file but that's actually panic prone
(storage asserts, on deinit, that there are no next ticks scheduled),
and, as a general rule, shutdown code that runs only on one operating
system only during upgrade is bound to become buggy sooner or later.