Vortex: Clean up data files
I'm not really sure why we didn't do this already... it was littering in `.zig-cache/.../` with data files after every run.

If output directory is specified then we don't delete the tree.

If the workload fails I think we won't delete the data files, since we exit with `std.process.exit(1);`.
(I will probably refactor that at some point, since that seems awkward).