StateMachine: Return the timestamp. Change the return type and semantics of the `create_accounts` and `create_transfers` operations.
The previous `create_*` operations only returned errors, omitting success results. While this is a good design choice to reduce the number of bytes transmitted by the wire protocol, in many cases the application needs a hint about _when_ the event was committed by the TigerBeetle cluster. Without returning the timestamp, this would require another network trip to look up the just-created event by ID.

This commit changes the return type to include the `timestamp` of the processed event. It also removes the `index` field, since now each event, both `ok` and errors, must have a corresponding result at the event’s index.

For successfully created events, the returned timestamp is the cluster timestamp used as a unique identifier. For imported events, it is the user-defined timestamp.
For failed events, the returned timestamp reflects _when_ the validation happened, allowing users to run point-in-time queries to diagnose, for example, what the balance was at that time or which transfers had taken place so far.
In some cases, the returned timestamp may be zero, because the event was not processed by the state machine (e.g., linked events within the chain _after_ the failed event, or in the edge case `linked_chain_open`).