metrics: handle sparse and non-zero starting enums more efficiently
Previously, we'd reserve space for an enum up to max(enum_tag). This was quite
inefficient for the `Operation` enum, since it starts at a relatively high
value and is used quite often.

Make it so that we store a compacted representation, more to reduce the worst
case packet count, but there's a tiny memory saving too :)