lsm: remove hash_log_mode support
HashLogMode is a debugging tool that jamii used once to debug a pretty
bad bug, but we haven't used it since. It feels better to remove the
code until we need it again, if ever. I also have a hunch that a lot of
determinism issues can now be debugged with `tigerbeetle inspect`.

The specific reason I want to remove this though is that I just want to
remove `config = @import("config.zig")` from build.zig, which makes
build.zig depend on the _entirety_ of tigerbeetle, requiring to
recompile the build script itself after every modification.