Vortex: Remove tracing
Previously vortex created a trace file with the various events it injects into the cluster.

It is a bunch of extra code and I think for this purpose the logs are superior anyway, since they interleave the different replicas as well. And for vortex in cfo I don't want to deal with multiple files, just one for the logs.