[StateMachine] Return the timestamp. Change the return type and semantics of the `create_accounts` and `create_transfers` operations.
The previous `create_*` operations returned sparse results containing only errors, omitting success results.
While this is a good design choice to reduce the number of bytes transmitted by the wire protocol, in many cases the application needs to know _when_ the event was committed by the TigerBeetle cluster. Without returning the timestamp, this would require another network trip to look up the just-created event by ID.

This commit changes the return type to include the `timestamp` of the processed event. It also removes the `index` field, since now each event, both `ok` and errors, must have a corresponding result at the event’s index.

For successfully created events, the returned timestamp is the cluster timestamp used as a unique identifier.
For imported events, it is the user-defined timestamp.
For `exists` events,  it returns the timestamp of the original event.
For other failed events, the returned timestamp reflects _when_ the validation happened, allowing users to run point-in-time queries to diagnose, for example, what the balance was at that time or which transfers had taken place so far.

The state machine still supports the previous behavior for old clients.