Vortex: Depend on TB executable in build.zig
This gives Vortex a default binary to test against -- a build of the current repo -- which is almost always what you want.

(A downside to this approach is that the running the integration tests now compiled TB an extra time.)

In the future this build will be a multiversion executable, so that Vortex can test upgrades.

I considered building from within `supervisor.zig` (rather than `build.zig`), since that would give us the ability to customize it at "runtime".
I'm still on the fence about this but decided to err on the side of keeping compilation in `build.zig`.