Vortex: Add seed argument
Vortex is non-deterministic. But the seed used still constrains things somewhat, since it controls fault injection. So if a test fails on a particular seed, it won't necessarily reproduce on re-runs -- but rerunning with the same seed probably finds that failure more frequently than a uniformly-random seed would.

Also, this way we don't need to special-case Vortex in CFO to avoid passing a seed.