VSR: Reserve 65535.x.x for cluster=0
The `65535.x.x` releases are now reserved for `cluster=0`.

This way, when testing multiversion binaries (either manually or with the integration tests' or (in the near future) Vortex's build) it isn't possible to use the test's multiversion build to upgrade a production cluster to non-production code.

(The reason that our existing development release `0.0.1` is not sufficient for these use cases is that to test upgrades we need a development version which is _higher_ than the real version.)