bus: separate message bus initialization from opening the socket.
The direct motivation here is that, without this change, tb_context.zig
fails to compile. The reason fro that is that the return type of `init`
(`!void`) differs between clients and replicas. Clients don't initialize
the listening  socket, and don't have associated errors.

Now that we have erased the type level distinction between the two,
tb_context.zig has to deal with this error types, even if they don't
arise dynamically.

I considered two ways to fix it:

* split `init_client` and `init_replica` methods
* split `listen` as separate from `init`

Both feel fine, and there's an argument that maybe we actually need to
implement both regardless, but given that _today_ signatures of inits
are the same, I decided to punt on that part for now.

And I like a separate listen, as that's a very distinct IO operation,
which is visible by the outside world, so its helpful to be explicit
about when we are ready to talk to someone!

Long term, I actually want to pull _initialization_ of message bus up to
main, and pass in the  Bus (rather than BusOptions) down to replica &
client, but that's for some future work!