bus: don't codegen accept path into clients
What a puzzling diff it is! Here's a clue:

    error: lld-link: undefined symbol: AcceptEx
        note: referenced by D:\a\tigerbeetle\tigerbeetle\src\io\windows.zig:377
        note:               tb_client.lib(tb_client.lib.obj):(io.windows.IO.accept__anon_39308__struct_41077.do_operation__anon_41973)

So what happens here is that most of WinSock functionality we use is
available in ws2_32.dll, which we link to. However, on accept path we
use `ws2_32.AcceptEx` function which, despite its Zig name, actually
comes from mswsock.dll. A direct fix would be to add dependency on that
.dll to our clients, but I think we have a different problem here.

The clients _shouldn't_ have accept path at all! While it'll never be
called, just having the code generated there in the binary is
suboptimal! Luckily, we can rely on Zig's lazy compilation semantics to
not generate calls to accept, if it is statically obvious that accept is
not called. Which we can do by separating ticks for client and the
replica!

I don't wan't to be super pedantic here, and I want to allow replicas
and clients to share the same tick, if they end up being in the same
process, so I don't forbid calling the general tick from the client, and
just add a specialization (calling specialized tick from replica would
be a mistake, and that is asserted).