tidy: re-implement function length check
This commit changes the implementation and the logic behind the check
for our infamous function length rule.

I tried to build the histogram of functions length using original logic,
and surprisingly learned that TigerBeetle is just 200 functions! I
wasn't able to debug the old code, so I re-implemented a much shorter
version:

* Zig guarantees that functions are sorted by their first token.
* Which makes checking if a function has any nested sub-functions just
  check the next function.

Then, I inverted the ratcheting logic. We originally tried to attack the
longest functions, but all the longest functions are actually boring,
and fixing them isn't satisfactory (and sometimes leads to worse code).

So instead the new idea is to define a red zone and say that functions
_just above_ the 70 limit are bad, but that, if a function is
sufficiently longer, it is good again. My hypothesis is that this will
lead to meaningful improvement, faster, and also addresses the main
problem: functions slowly growing larger over time and boiling over the
70 limit.

The current red zone is tiny though, just `71`, though I think it
already revealed an opportunity for a meaningful simplification! In Scans,
rather than writing

            if (probe_key < self.key_min) {
                if (self.direction == .descending) self.abort();
                return;
            } else if (self.key_max < probe_key) {
                if (self.direction == .ascending) self.abort();
                return;
            }

, it makes sense to semantically compress switching on direction:

    if (self.direction.cmp(probe_key, ."<", self.key_lower)) return;
    if (self.direction.cmp(self.key_upper, ."<", probe_key)) {
        self.abort();
        return;
    }

Though, that is for future me to deal with. For now, I split the `probe`
function into fast path with early returns, and `probe_advance` which
actually is guaranteed to be a non no-op.