lsm: increase the level of abstraction when handling directions
Tree LSM is a sorted array with a monocle and a top hat.

We want to iterate it in both directions:
- For CDC, you want to learn about all new objects with timestamp>thershold.
- For paginated timelines, you want to learn about past objects with timestamp<thersold.

Sadly, this can't be implemented via just a single branch near the end of the system, we need to
change a whole bunch of `<` to `>` throughout the stack.

Direction encapsulate the logic of "if ascending use < if descending use >". The mnemonic is
that usual comparison is horizontal along a number line, but Direciton-aware is vertical.

In other words, `key_min` and `key_max` track natural ordering, while `key_lower` and
`key_upper` are direction-aware.