bench: implement micro-benchmarking harness
We have a bunch of benchmarks here and there, but they are somewhat
bitrotted. The big idea behind this module is to fix that by providing
two sets of parameters:

- one to run the benchmark quickly during the tests,
- the other to the benchmark "for real"

I've only updated the binary search benchmark as a proof of concept, but
it would be a great exercise to go through our existing benchmarks and
refresh them!