aof: stat the aof on checkpoint and reopen if not there
This adds the ability to rotate AOF files, without having to use signals.
Instead, you can `mv` the old AOF out of the way, and tigerbeetle will
continue writing to it until checkpoint, upon which it'll notice and
reopen the original path.

The main reason behind this is to allow trimming AOF sizes without having
to interrupt the cluster.

When the cluster hits a checkpoint, it'll `stat` the path, see there's no file
there anymore, and reopen a new fd, giving a clean rotation.

The _safe_ way to rotate and backup is thus:
* `mv current.aof old.aof`
* # Wait for current.aof to be recreated, this is NB!
* `s3 cp old.aof s3://my-backup/`

The middle step is load bearing! Before it happens, tigerbeetle is still writing
in to the same file descriptor (which now points to old.aof), so old.aof _is not_
safe to copy or delete.

There's no way to rotate other than waiting for a checkpoint - this is to keep
things simple. If it's a very lightly loaded cluster, you can pulse no-ops or
similar through to force one.